[exporter/awss3exporter] Add middleware to recalculate AWS V4 signature for GCS endpoint
This update introduces a new middleware, `RecalculateV4Signature`, which recalculates the AWS V4 signature after modifying the `Accept-Encoding` header.

This middleware is applied when the S3 uploader is configured to use the Google Cloud Storage endpoint. The change ensures proper signing of requests to GCS, enhancing compatibility with AWS SDK's signing requirements.